requirements-hytrel.txt
# pip==24.0
pyarrow==8.0.0
# scipy==1.12.0
# torch==1.13.0
# transformers==4.21.1
transformers==4.45.2
tokenizers==0.20.1
pytorch-lightning==1.7.7
# lightning==2.4.0
# torch-geometric==2.6.1
torchmetrics==0.10.0
torch-geometric==2.0.4
torch-scatter==2.1.1
torch-sparse==0.6.18
deepspeed==0.9.5
ujson==5.8.0
datasketch==1.6.5
faiss-cpu==1.9.0
loguru==0.7.2
numpy==1.26.4
matplotlib==3.9.2
scikit-learn==1.5.2
hdbscan==0.8.39
seaborn==0.13.2
openpyxl==3.1.5